Add a data directory to the tree. This is the default per-user data that
* data/.cvsignore,
data/Makefile.am:
Add a data directory to the tree.  This is the default per-user
data that nautilus installs the first time it runs.

*data/top/.cvsignore,
data/top/Makefile.am,
data/top/Welcome:
Add a top directory to the tree.  The top directory is an
alternative home place for novice users.

*configure.in,
Makefile.am:
Grok the above.