Add support for keeping track of changes within a whole preference
* nautilus-widgets/nautilus-preferences.c:
(pref_hash_node_add_callback), (pref_hash_node_remove_callback),
(preferences_gconf_callback):

Add support for keeping track of changes within a whole preference
namespace.

Pass in the expected preference to the gconf callback.  Do this so
that we can compare the expected prefrence name with the one that
gconf reports.  If they dont match, then we have something that
changed within a namespace and not an individual key.

Also fix some memory leaks.  Was forgetting to free a couple of
returned preference names.