Make location bar more reactive: path bar, toggle button and zoom control
The buttons in the directory button bar are made inactive when the
corresponding pane is inactive. This means that the buttons have an
insensitive appearance, but are in fact clickable. Clicking on them
will make the corresponding pane active while changing into the
respective directory.

This required the pathbar to export the button widget, done with the
new function nautilus_path_bar_get_button_from_button_list_entry().