Draw background of inactive panes in INSENSITIVE color.
In order to make it easier to visually distinguish active and inactive
panes, the background of inactive panes is drawn in the theme's
INSENSITIVE color. Like the inactive location bar widgets, this is
only a visual marker, and does not mean that that pane is
insensitive. It's still clickable, and in fact a click makes the
corresponding pane active.