Centralize extra view ref and sink if required
Doing this in only one place makes sense as its clearer.
Previously we ref:ed one extra time for the window->view ref but left
the floating ref to be taked when adding the widget to a parent.
This breaks when we free the view before adding it to a parent, so
we need to properly sink it.