Don't load files for thumbnailing that we can't thumbnail anyway
When zooming out we read the real image and use that as the thumbnail.
However, we should only do this for files gdk-pixbuf can read as other
things can't be read anyway and may potentially be very large (such as
videos) since the thumbnail size limit only applies for images.