Move slots from NautilusWindow to NautilusWindowPane
A window does not hold lists of slots anymore, but lists of panes, which
in turn hold lists of slots, making a slot effectively a layer between
a window and a slot. Much of the remaining part of the patch is
simply replacing "window->details->active_slot" with
"window->details->active_pane->active_slot" and similar changes. Commits
to move corresponding parts of the logic out of the window into the pane
are to follow.