Better split view size allocation
Ensure that the views always get *some* size so that we can at least
always see the scrollbars, etc.

Also make the toplevel size of a pane be a fixed size, thus guaranteeing
the two panes in a split view return the same size request, thereby
guaranteeing a default (if unset by user) position of 50%. Without this
we could get e.g. a size request from the tabs or the location bar size
change causing a change in the view splitter position.