Better tab size
This removes the minimum size of tabs, since that causes problems
when there are little availible horizontal size (such as in split
view), as it caused ellipsizion to only work for strings larger
than the minimum set size.

However, doing *just* this causes the tabs by default to be only
3 chars wide, which is obviously also bad. So we enable tab-expansio
to solve this.

The new situation is not ideal either, but its the best we can do
before this gtk bug is fixed:

https://bugzilla.gnome.org/show_bug.cgi?id=125250