Fix window slot close crash
On window close via nautilus_window_close when the viewed file
closed there was a ref on the slot living past the window. This
caused crashes when disposing the slot as it referenced the window.
We fix this by running the dispose handler when closing a slot
and by making the dispose handler able to run multiple times.