Remove switch to other pane menu item and make F6 do this instead
This is a very useless menu item, its sole reason for existance is
to bind a keyboard combo to it. Instead we should just choose one.

Some research indicate that F6 is standard for this. In Gtk+ F6 is
used for split view navigation in general, so it will toggle between the
sidebar and the two views. However, you can reach the sidebar in other
ways using keynav, and the more important thing is the pane toggling,
so we override this and make F6 be pane switching only.