Remove tabs menu
I don't think this menu is much use, and it clutters up the menus even for
users who don't use tabs, which is against our policy of having the more
advanced features affect users not using them as little as possible.

Also, remove the setting to disable tabs althoughter, as with the menu
removed the tabs feature is very minimal when not in use.