Set exit_with_last_window to false by default
This fixes problems with show_desktop=false causing nautilus to
exit and gnome-session restarting nautilus in a loop (#571417).

This means nautilus will keep running, so that it can handle automount
and stuff. If you only run nautilus rarely you need to set this to true.

In the future we will move the automount handling to gnome-settings-daemon
and then we will probably change this back.