Add possibility to register an external bulk rename tool
If the new setting "bulk-rename-tool" in org.gnome.nautilus.preferences
is set to a non-empty string, it is treated as a bulk renamer and
gets invoked on a Rename action if multiple files are selected.
Nautilus appends a space separated list of URIs of all selected
files to the command string.

https://bugzilla.gnome.org/show_bug.cgi?id=306489