smclient: remove XSMP support
EggSmClient is not well-maintained, and it doesn't seem like it will
ever hit GTK+, and the overall implementation seems to be poorly tested
and buggy.

Moreover, moving to 3.0 with nautilus being an application, it makes
less sense to have it save opened windows state, as it might not be
always running at all.