Rework architecture of NautilusWindow/NautilusWindowSlotInfo
In the past, nautilus was designed keeping in mind the concept that it
might not just be a file browser, but a sort of platform to embed
various kind of views in it.
  - libnautilus-private: general lowlevel items, such as the icon
    container, NautilusFile, I/O and so on
  - src: general widgets and UI items
  - src/file-manager: file manager views, agnostic of what's in src/,
    for the sake of being pluggable enough

This is really not applicable anymore, as nautilus won't go in any
direction other than being a file manager; at the same time, this
complicates code and architecture quite a lot.

In this commit, as a first step towards a more clean codebase, we remove
the NautilusWindowInfo and NautilusWindowSlotInfo interfaces, using
NautilusWindow and NautilusWindowSlot directly.

Note that there should be no actual code changes, only moving of pieces
around.