navigation-window: initialize actions before setting up the pane
This is needed so that the pane can use those actions to build the
toolbar. We also have to remove the code that updates split view
actions from _initialize_actions(), as it requires the pane is realized.
It should not be much of an issue, as that code is called when loading
an URI anyway (and so it is called afterwards when loading the window).