Fix up eject button hover in places sidebar
We were calling gtk_tree_view_column_cell_get_position() without
properly loading the cell attribute for the right row before.
We fix this by calling gtk_tree_view_column_cell_set_cell_data().

With this in place we can also use the x_offset for the position and
avoid the whole summing of widths.

Due to a bug in Gtk which expands the eject icon cell renderer we
have to right align it so that it lines up properly.

https://bugzilla.gnome.org/show_bug.cgi?id=640741