Use ctrl-delete as the keyboard shortcut to trash files
This change was made to make it harder to accidentally trigger a file
delete. We still support the trash that will let you get a trashed file
back, and we will get undo support to make this even easier. However,
that only works if you know you deleted the wrong file, not if you
accidentally hit delete while the nautilus window was focused.