dnd: rework handling of _NETSCAPE_URL dnd links
Previously we used to to the following:
* default to always "ask", unless the copy was explicit
* when dropped, we did not actually ask anything but we would trigger an
  async mimetype query_info and:
  - trigger a file asking whether to download or to link for text files
  - silently link for HTML files
  - silently download for every other mimetype
* if the query_info was not completed within one second, we would
  download by default

This is somewhat broken, as the DnD icon is not what you would expect;
also downloading is not usually done with DnD from the browser, and it's
also expensive, so it should never be the default, but an explicit
choice. It also makes it impossible to create links for anything else
than text or HTML.

Change the policy to always link by default, unless the user explicitly
requests the copy.