application: don't look at DESKTOP_AUTOSTART_ID anymore
Now that we don't support being a session component anymore, it doesn't
make sense to special case the session autostart. This also breaks
opening new windows when nautilus is added to ~/.config/autostart, which
is now the supported way of having desktop icons automatically, as we
end up reading the env variable every time when parsing the command line
(as the original process will always have that set) and refuse to show a
window.

https://bugzilla.gnome.org/show_bug.cgi?id=649063