window: reverse the order of key-press event processing
Usually the default GtkWindow handler for key-press events processes
them in the following order:
- calls gtk_window_activate_key() to process mnemonics/accelerators for
  the toplevel window
- calls gtk_window_propagate_key_event() to propagate the events to the
  focus widget
- chains up to parent if both fail

We want gtk_window_propagate_key_event() to be called before
gtk_window_activate_key(), as when we're focusing an editable widget
(e.g. renaming a file), we want all keybindings to apply to that, e.g.
Delete, Ctrl+Delete (or Ctrl+W if we're e.g. using an emacs-mode
GTK+ binding set). This interferes a bit with the type-ahead search
windows that NautilusIconContainer and GtkTreeView pop up; we can
control the former, but not the latter, so we need a bit of a hack in
NautilusListView to prevent the search window to steal the pasted URI,
and still not handle the event.

https://bugzilla.gnome.org/show_bug.cgi?id=314431