application: rewrite GApplication command line handling
Do the command line parsing in local_command_line for the set of options
we're interested in parsing there (e.g. directories to open, self checks
and --quit); move the rest of initialization in
nautilus_application_startup() and use g_application_open() and a custom
"quit" action to forward the requests to the main application.

This also fixes these two bugs.

https://bugzilla.gnome.org/show_bug.cgi?id=649336
https://bugzilla.gnome.org/show_bug.cgi?id=637157