desktop: ensure the desktop metadata is up to date when setting as ready
As part of the I/O machinery cycle internal to call_when_ready
nautilus_file_clear_info() can be called and clear the metadata we
previously set during _init.
Move the desktop metadata initialization before returning the file as
ready, so it's always up to date when the view reads the settings.

https://bugzilla.gnome.org/show_bug.cgi?id=655561