window-pane: fix switching to the other pane when closing the last slot
We're closing a window slot; when it's the last slot to be closed, we
used to call nautilus_window_split_view_off() on the wrong pane.
Even if we called it on the right pane, it wouldn't have worked anyway,
because we could have removed the active pane, and that function always
assumes going back to it.
Also, we need to ensure that active_slot is set on the new pane, or
nautilus_window_set_active_slot() won't be called to restore the right
window state.

https://bugzilla.gnome.org/show_bug.cgi?id=652320