places-sidebar: don't call gtk_tree_model_get() with an invalid path
If we click on an empty area of the sidebar, we shouldn't pop up any
menu at all anyway, which is what this commit changes.
Previously, we would have tried to get a GtkTreePath for the currently
selected point and an iter for that (invalid) path, which would lead to
gtk_tree_model_get() segfaulting on an invalid iter.

https://bugzilla.gnome.org/show_bug.cgi?id=662218