window: explicitly destroy the sidebar in _destroy()
When the nautilus window is destroyed, make sure to explicitly destroy
the sidebar before the panes.
Destroying all the panes can trigger a callback back into the sidebar,
which would then try to access the window pane list from the callback.
The pane list was already destroyed and cleared at that point though,
and we would segfault.

https://bugzilla.gnome.org/show_bug.cgi?id=652320