file-undo-manager: use NautilusFileUndoInfo
Rework NautilusFileUndoManager to use the new NautilusFileUndoInfo
objects.
At the same time, make it use a single item instead of a queue, turning
it into a state machine. This has various reasons; from a code
perspective, if we keep a stack of operations, we should also keep track
and invalidate items in there when something changes on the filesystem,
but this can be very expensive.
On the other hand in the UI we probably don't want to expose more than
one item anyway.