places-sidebar: change the way we restore selection on update
Instead of checking whether the URI of every item we add is a candidate
for selection, cycle through all the added items after inserting.
This also fixes a bug where the GtkTreeIter we used for keeping track of
the last added item was not up-to-date with the actual cell, since the
sort function we apply to the items will modify the position of the item
in the tree itself, making the iter invalid.

https://bugzilla.gnome.org/show_bug.cgi?id=668212