file-operations: use the right prototype for empty trash callback
When unmounting, we might first prompt whether the trash should be
emptied or not on the target volume. If we do so, make sure the callback
from the empty trash job has the right prototype, to avoid crashing
later trying to interpret a gboolean as a pointer to a data structure.

https://bugzilla.gnome.org/show_bug.cgi?id=670595