a11y: add a method to create a derived GtkWidgetAccessible type
For EelEditableLabel, we need to create an object derived from the
accessible type of the parent class; since that will be a
GtkWidgetAccessible, which is private, we can't just subclass it using
the regular GObject mechanisms.
This functions creates a derived GType by peeking at the parent
accessible class and type.

https://bugzilla.gnome.org/show_bug.cgi?id=672498