view: don't connect to NautilusWindow signals
Since the window of the slot can change, we don't want to connect
signals from the view to the window.

We only connect to NautilusWindow signals for two reasons:
- updating the shadow type of the content view when the number of tabs
  in the notebook changes
- following the hidden-files-changed signal

Since the first is really a hack, it can live in NautilusWindow itself
To implement the second, move the Show Hidden Files action to the view
action group, and handle hidden files action changes completely in the
view.