desktop-canvas: set margins according to the primary workarea
This ensures desktop icons are pinned to the primary monitor, instead of
moving around when plugging in a new display.
There's also a slight change in behavior introduced by this patch - that
is for multihead configurations where the first monitor in visible order
is not the primary one, it won't be possible to move desktop icons to
monitors to the left, or to the top, of the primary monitor - in other
words, desktop icons layout starts from the primary monitor and expands
rightwards/downwards.

https://bugzilla.gnome.org/show_bug.cgi?id=420624