mime-actions: resolve symbolic links before launching the application
We already do this for all other kind of references, and I think it
makes sense to do it for symbolic links too.
This also ensures applications don't have to deal with e.g. ensuring
backup files are saved relative to the original file and not the
symlink, as reported in

https://bugzilla.gnome.org/show_bug.cgi?id=686465