slot: ensure we're back from an active search when changing location
We used to keep track of the search active state using the state of the
search toggle button; unfortunately, the button is per-window,
so its active state will also be synchronized when switching tabs.
This will cause a non-search slot that becomes active following a tab
switch from a search slot to think its search just became inactive, and
will then change its location to the default ($HOME).

Fix the bug by keeping an additional state per-slot for whether there's
an active search.

https://bugzilla.gnome.org/show_bug.cgi?id=686893