window: don't remember focus widget when switching search visible
The workaround we need to make to focus the entry on-the-fly and
synchronizing the visible state with the toolbar button badly clash with
the focus widget state saving mechanism we have in place.
We end up calling remember_focus_widget() on the newly-focused
search entry in certain circumstances, so we end up restoring the focus
on the (then hidden) search entry, which will then eat key events while
invisible.

Sidestep the issue entirely by just focusing back the view when closing
the query editor.