query-editor: Don't set the entry's text if not strictly needed
If the current text in the entry is equal to the new query's text we
shouldn't needlessly set the text on the entry since that affects the
cursor position.

gtk_entry_set_text() already does a similar check internally but that
isn't enough here because NautilusQuery doesn't keep the full string
but instead a stripped version of it.

https://bugzilla.gnome.org/show_bug.cgi?id=696430