query-editor: Fix initial key press processing with async IMs
When using an async input method we can't rely on the entry's text or
preedit text being updated as a direct result of processing a key
event.

Instead, we can look at the event keyval and check if it's going to
yield a printable character to decide if we are interested in the
event.

https://bugzilla.gnome.org/show_bug.cgi?id=696260