search-provider: Return TRUE from GDBus method handlers
As Matthias Clasen noticed, the signal handlers for the gdbus-codegen
skeleton for the shell search provider have the wrong signature, and
should return TRUE to indicate that the signal has been handled.
Otherwise, it is down to chance if the method is handled or if a
method-not-implemented error is emitted. It seems that the GCC option
-fstack-protector-strong causes the default return value to be
interpreted as FALSE, so this might explain why the problem was only
noticed by a wide variety of users recently.

The bug manifests as Nautilus not returning any search results when
using the gnome-shell activities overview search.

https://bugzilla.gnome.org/show_bug.cgi?id=692041