file-operations: Don't crash after failed inhibit
If gtk_application_inhibit fails, it returns a 0 cookie.  When this
invalid value is passed to gtk_application_uninhibit, Nautilus crashes
if glib or gtk is compiled with minimal debug checks.
So, only call gtk_application_uninhibit if a non-zero cookie is
returned.  Also, use the correct type for inhibit_cookie (guint).

https://bugzilla.gnome.org/show_bug.cgi?id=709015