monitor: improve unmount detect heuristic
Detect when an unmount happens and a subfolder of the mount is
displayed. The previous solution did a GMount pointer compare, but this
was not working in any case.
This patch fixes is by comparing the removed mount root with the
currently monitored location.
This fixes a regression since 0852847e2613dbcebb4ed7f58e5b2aee3aa86a90.

https://bugzilla.gnome.org/show_bug.cgi?id=724560