application: Add missing unref for NautilusFile used in network connections.
This was introduced in commit 95d42ea01f7f9c60210b415554e7bbf6f4e1b334 and released in 3.5.91.
The error meant that when a network location was disconnected the next time you connected to it you would get the old NautilusFile object that was marked as "gone" and hit an assertion.

https://bugzilla.gnome.org/show_bug.cgi?id=708282