Update window title when a search string is updated
nautilus_window_slot_set_location updates window slot
title only when a directory is changed. This is OK
when going through the directory structure. But when
a search is started, "search" directory is created
with some name (create_new_search) and this name
is updated as the search is updated (sync_search_directory)
- but the location instance remains the same
until the search is finished. Thus we need to update
the title even if the location remains the same.

https://bugzilla.gnome.org/show_bug.cgi?id=698767