NautilusPropertyDialog: Fix problems with fast group/owner changes
If the owner or group of a file was changed twice very quicky then
the second change would cancel the first. However, the code assumed
that after the nautilus_file_cancel() call the result of the first
change would not be called. That is not true anymore, and when
it did get called it tried to stop the timed wait, which gave a warning,
and in general touched stuff that the new operation should only touch.

We fix this by making A group/owner change its own struct which we
delay freeing until the operation has actually finished, although we
track if it has been cancelled so we can avoid multiple calls to
eel_timed_wait_stop().