file: Let renames work on Google Drive
Renaming an item on Google Drive does not change the actual GVfs path
because the path is made up of document-ids, which do not change. A
rename operation only affects the title of the entry. In GIO terms,
only the standard::display-name changes, but the standard::name remains
the same.

Therefore, if there is already a file with the same standard::name,
then don't mark it as gone if it is the same file that we are renaming.

https://bugzilla.gnome.org/show_bug.cgi?id=740383