nautilus-view: drop empty trash on view
When the user is in the trash, there's already a inapp dialog that shows
a button to empty the trash. On the other hand the sidebar also shows
this item.
Since it's not expected the user to see the trash directly as a folder,
drop this item from the menus.