general: rework menus of nautilus
Refresh the nautilus menus in the view (context menus) and the menus in
the toolbar, changin as well in the path the UI of the toolbar to match
mockups.

In the way, port to GAction and GMenu every action and menu of nautilus
and some clean ups thanks to this port.