nautilus-floating-bar: hide on hover
Due to the floating bar being in an overlay, it can obscure the content
under it.

We were planning to remove it and use an action bar. But it's taking
long, so in the meantime we can improve this situations hiding the
floating bar when hovered.

On the way I improved the handling of the spinner, which was failing
to be shown on certain situations.

Patch based on Nelson Benitez, thanks!

https://bugzilla.gnome.org/show_bug.cgi?id=651293