general: Implement extensions for GAction
The last commit rework the menus and ported to GAction.
Update the extension mechanism to that.

The changes are very little. Now menu items in extensions lost
the property 'priority' since gmenu doesn't have one, and gain
two properties, menu_name, which is in which menu do you want the
extension to add a menu item, it can be background or selection, and
menu_section, which indicates in which section of the menu do you want
to add the items of the extension.