nautilus-file-undo-operations: Fix crash for multiple deleted files
Previously when using GtkMenuItem, setting NULL to a label of a menu
item was not a problem, although it was a bug anyway since the
label was present as empty.
We were doing that with the Undo and Redo labels on the toolbar.

Now with GMenuItem it was crashing when deleting more than a file since
the label was NULL and the menu item couldn't be created.

To fix it, make sure we always set a label for the trash delete job.