nautilus-view: Change move to trash accelerator
Now that we have an in app notification, the user has a feedback if
they push Delete key as an accident. So we no longer need to make
the move to trash action difficult to do.
So change the accelerator of move to trash from <shift>Delete to
just Delete

https://bugzilla.gnome.org/show_bug.cgi?id=743630