general: Add in app notification for move to trash
We want to make sure the user has feedback when moving files to trash,
so add an in app notification when some file has been moved to trash.

Currently the notification manager only supports one notification at
once, since the undo machinery only support to undo the last operation,
not a specific operation.
Even if we add support to it I'm not sure if multiple notifications of
deleted files would be good, so instead of that just show one at once.

To achieve that, connect to the undo state and delete all notification
when the undo state changes and create the new notification if the undo
operation is "move to trash" and if the undo manager state is "undo".

https://bugzilla.gnome.org/show_bug.cgi?id=743630