nautilus-view: Don't update if not active or destroyed
For context menus we update menus on a timeout. For the cases of
the toolbar afaik that's not needed so we update directly.
The schedulers for the context menu has a protection against updating
if the view is destroyed or not active.
Apply that protection to the update of the toolbar so we don't crash
if we call it while destroying.