nautilus-window-slot: Don't grab focus on view change
We were grabbing the focus on view mode change, so if the user changed
the view mode with keyboard he could continue using it.
But it makes the problem that with the new popover menu in the toolbar
where you can choose between different views while the menu is open,
grabbing the focus to another widget dismiss the menu, which is
against the behavior of the GtkMenuPopover.

To fix that, don't grab focus on view changes.

In the tests I have done, you can still focus the view just using the
arrows after the view changes, so I don't think it will be a problem for
keyboard users.

https://bugzilla.gnome.org/show_bug.cgi?id=743591