desktop: change "home" to "Home"
Modified "home-icon-name" so that the home folder is displayed as "Home"
when displayed on the desktop (in Classic mode or when turned on using
Tweak Tool). Currently, the text for the home folder is "home" with a
lowercase h, but "Trash" and "Network Servers" start with capital
letters. This change makes the home folder consistent with the other
desktop icons.

https://bugzilla.gnome.org/show_bug.cgi?id=744813