configure: enable debug always
We can control the debug output with G_MESSAGES_DEBUG anyway,
and is off by default.
Probably we were doing this to avoid the compilation and performance
impact of the debug stuff, but that shouldn't be a real problem these
days.
This also allows to the final user run the application with debug
without the need of building the application.