nautilus-file: explain why is not possible to rename
desktop files are managed as a special case, and changes on the file
name also affect the contents. But when the content are not a desktop
valid format, it generates an error.

The only thing that comes to my mind to keep maintaining this special
handling of desktop files and also not being in a inconsistent state is
to just explain better why we failed to rename.

https://bugzilla.gnome.org/show_bug.cgi?id=721411

https://bugzilla.gnome.org/show_bug.cgi?id=734401