nautilus-view: allow using delete shortcut in the trash
Until now the "move-to-trash action" was disabled while in the trash,
and only the "delete" action was enabled. This had the problem that
that action was used for the "delete permanently" menu item, which its
visibility relies on a setting and uses a different shortcut than
move-to-trash.

To allow the user to use the delete shortcut in the trash as well
create a new action so we can control the visibility and the shortcut
differently than the Delete Permanently and rename the actions to make
it more clear.

This is a fallout of the menus rework, from the clean up of the
actions.