nautilus-view: always allow to permanently delete
Before 3.15 we had a setting for adding a menu item for permanent
deletion of files. But the shortcut was always enabled no matter the
state of the setting, which seems inconsistent.

Then with the GAction port we changed that behavior to be consistent,
but people that were unaware of the setting were complaining that
the shortcut for permanent deletion was not working.

Given that we show a dialog if we hit that shortcut, and the setting
itself is confusing, just enable always the shortcut but remove the
setting and the menu item for it.

https://bugzilla.gnome.org/show_bug.cgi?id=745321