nautilus-window: don't use libgd notifications
We are using notifications just for one window and one type of
notification and only one notification at once, so we don't need to use
notification-manager from libgd.
Given that the manager is not necessary, the reasons to use libgd
notifications becomes weak.

Instead of libgd notifications create our own with a GtkRevealer and
remove all the libgd notifications, so we have a cleaner code.