query-editor: cancel search with entry's signal
After the introduction of GtkSearchEntry by commit
5521f24f, the Esc key stopped working, specifically
after the introduction of the GtkSearchEntry::'stop-search'
signal.

The old approach of adding a GtkBindingSet directly to the
NautilusQueryEditor class then stopped working, as the Esc
key won't be propagated by the search entry. Without the key
catch, the NautlisQueryEditor::'cancel' signal couldn't be
fired, which causes the Esc issue.

By listening to the GtkSearchEntry::'stop-search' signal
and sending NautilusQueryEditor::'cancel' right after,
the Esc key is correctly detected and the problem is
solved.

https://bugzilla.gnome.org/show_bug.cgi?id=745305