Register double-click on button release
Only perform the double-click action when the second click is released.
This fixes a strange problem where double-clicking on a
password-protected webdav share (as exported by gnome-user-share) fails
to show the password dialog (which gnome-shell does) due to Nautilus
holding a grab. By only taking action when the click is released,
Nautilus is not holding a grab and the password dialog is shown.

When it fails, the following is shown in the log:
gnome-session[20151]: Gjs-Message: JS LOG: pushModal: invocation of begin_modal failed

https://bugzilla.gnome.org/show_bug.cgi?id=745859