desktop-canvas-view: reorder if icon size changes
We changed the icons zoom levels icon sizes, which works fine on
the common nautilus window since the icons are positioned every time
we create a new view, etc.

The desktop view save the icon positions, so when we changed to a
larger size of icons by default the icons overlapped each one, looking
really bad for the people who update from an older version of nautilus.
This doesn't happen in clean installs, since there is not saved
positions of the icons.

To fix this rearrange the icons every time we found that the icons were
stored with a icon size different than what we have now or if we don't
have this metadata available.
That will only happen the first time the user runs nautilus with this
patch, since that key of the metadata won't be available; or every time
we change the hardcoded icon sizes for the zoom levels.

https://bugzilla.gnome.org/show_bug.cgi?id=745669