window: don't always make sidebar visible on new windows
We are setting the sidebar as visible in the ui definition and binding
this visibility to the disabled-chrome property (inversely) to prevent
the sidebar from being visible in the desktop. But this also makes
the sidebar visible in all new windows, ignoring both the state of the
"Sidebar" toggle in the app menu and the start-with-sidebar setting.

Fix this by setting the sidebar as not visible in the ui definition and
removing the binding. The sidebar will be made visible if appropriate
in the nautilus-window-initialize-actions function.

Fixes: https://bugzilla.gnome.org/show_bug.cgi?id=746217