nautilus-application: don't create desktop on Wayland
We were doing some calls to the x11 backend, which are not supported on Wayland
(or another backend diferent than X11) and the application were crashing when
creating the desktop on Wayland.

To workaround it, simply don't create the desktop if we are on
a different backend than X11.

Future work will be to split the main application and the desktop part.

https://bugzilla.gnome.org/show_bug.cgi?id=746286