view: show "New Folder" dialog
This commit introduces the "New Folder"
dialog, which asks the folder name before
actually creating it.

With the introduced changes, the folder
is created with the given name instead
of creating it first with the generic
"Unamed folder" and then renaming it.

The dialog has a "delayed message" logic,
where typos are immediatly recognized, but
coincident folder names appear ~2s later.

This dialog is part of the ongoing effort
to modernize Nautilus to better fit GNOME
standards, and the latest mockups can
be found at [1].

[1] https://raw.githubusercontent.com/gnome-design-team/gnome-mockups/master/nautilus/nautilus-next/create-folder-wires.png

https://bugzilla.gnome.org/show_bug.cgi?id=747381