view: Make the zoom slider the preference itself
With the recent changes of the zoom, some people wanted to
have by default the smaller zoom level. They changed the slider
and expected to set the setting as well, so further applications
openings had that zoom level.

But changed on the zoom were intended only for the current view
and not altering the preference. And instead, if you wanted to
make it permanent you needed to go to Preferences and change it there.

Seems that was confusing for the users, so instead use the slider
as the permanent preference and remove the zoom preferences from
the Preferences dialog.